requirements.txt
transformers>=4.30.0
torch>=2.0.0
datasets>=2.12.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
groq>=0.3.0
langchain>=0.0.267
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2 